+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 ## Pending
 ### Added
 - `CountdownEvent::wait_timeout`, to wait but also have a timeout
+- `SignalEvent`, a primitive that lets one or more threads wait for a signal from another one
 
 ## [0.1.0] - 2017-02-27
 ### Added
-- CountdownEvent, a primitive that keeps a counter that allows you to block until it hits zero
+- `CountdownEvent`, a primitive that keeps a counter that allows you to block until it hits zero
 
 <!-- vim: set tw=100 expandtab: -->